Bonus-2

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: nettalha <nettalha@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/27 10:56:52 by nettalha          #+#    #+#             */
-/*   Updated: 2022/10/27 11:00:29 by nettalha         ###   ########.fr       */
+/*   Updated: 2022/11/27 03:18:17 by nettalha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
